Moved test to file where ssl configs are tested
